RIP store-api
