+++ Cargo.toml
@@ -23,11 +23,12 @@ members = [
     "common/exception",
     "common/tracing",
     "common/profiling",
-    "common/store-api",
     "common/kv-apis/util",
     "common/kv-apis/api",
     "common/kv-apis/vo",
     "common/meta-apis/api",
+    "common/dfs-apis/api",
+    "common/dfs-apis/vo",
     "common/sled-store",
     "common/stoppable",
     "common/store-api",
