+++ common/dfs-apis/README.md
@@ -1,0 +1,10 @@
+**NOTE**
+
+currently, crates under this directory are for `RemoteTable` accessing data blocks, 
+which is NOT what DFS client/api are supposed to do.
+
+After `RemoteTable` is deprecated, crates here will be migrated into real dfs clients 
+
+
+
+
