+++ common/dfs-apis/vo/src/lib.rs
@@ -1,25 +1,26 @@
-// Copyright 2020 Datafuse Labs.
+//  Copyright 2021 Datafuse Labs.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
 //
 
 use common_datablocks::DataBlock;
-use common_datavalues::DataSchemaRef;
 use common_planners::Part;
 use common_planners::PlanNode;
-use common_planners::ScanPlan;
 use common_planners::Statistics;
-use common_streams::SendableDataBlockStream;
+
+// TODO A better name, we already have a SendableDataBlockStream
+pub type BlockStream =
+    std::pin::Pin<Box<dyn futures::stream::Stream<Item = DataBlock> + Sync + Send + 'static>>;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, Eq, PartialEq)]
 pub struct DataPartInfo {
@@ -89,39 +90,4 @@ pub struct AppendResult {
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, Eq, PartialEq)]
 pub struct TruncateTableResult {
     pub truncated_table_data_parts_count: usize,
-}
-
-// TODO A better name, we already have a SendableDataBlockStream
-pub type BlockStream =
-    std::pin::Pin<Box<dyn futures::stream::Stream<Item = DataBlock> + Sync + Send + 'static>>;
-
-#[async_trait::async_trait]
-pub trait StorageApi: Send + Sync {
-    async fn read_plan(
-        &self,
-        db_name: String,
-        tbl_name: String,
-        scan_plan: &ScanPlan,
-    ) -> common_exception::Result<ReadPlanResult>;
-
-    /// Get partition.
-    async fn read_partition(
-        &self,
-        schema: DataSchemaRef,
-        read_action: &ReadAction,
-    ) -> common_exception::Result<SendableDataBlockStream>;
-
-    async fn append_data(
-        &self,
-        db_name: String,
-        tbl_name: String,
-        scheme_ref: DataSchemaRef,
-        mut block_stream: BlockStream,
-    ) -> common_exception::Result<AppendResult>;
-
-    async fn truncate(
-        &self,
-        db: String,
-        table: String,
-    ) -> common_exception::Result<TruncateTableResult>;
 }
