+++ common/kv-apis/api/src/lib.rs
@@ -15,6 +15,6 @@
 mod kv_api;
 mod kv_api_sync;
 
-use common_store_api_util as util;
+use common_kv_api_util as util;
 pub use kv_api::KVApi;
 pub use kv_api_sync::SyncKVApi;
