+++ common/store-api-sdk/src/impl_flights/storage_api_impl.rs
@@ -26,17 +26,17 @@ use common_arrow::arrow_flight::utils::flight_data_to_arrow_batch;
 use common_arrow::arrow_flight::Ticket;
 use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
+pub use common_dfs_api::StorageApi;
+pub use common_dfs_api_vo::AppendResult;
+pub use common_dfs_api_vo::BlockStream;
+pub use common_dfs_api_vo::DataPartInfo;
+pub use common_dfs_api_vo::ReadAction;
+pub use common_dfs_api_vo::ReadPlanResult;
+pub use common_dfs_api_vo::TruncateTableResult;
 use common_exception::ErrorCode;
 use common_planners::PlanNode;
 use common_planners::ScanPlan;
 use common_runtime::tokio;
-pub use common_store_api::AppendResult;
-pub use common_store_api::BlockStream;
-pub use common_store_api::DataPartInfo;
-pub use common_store_api::ReadAction;
-pub use common_store_api::ReadPlanResult;
-pub use common_store_api::StorageApi;
-pub use common_store_api::TruncateTableResult;
 use common_streams::SendableDataBlockStream;
 use futures::SinkExt;
 use futures::StreamExt;
