+++ common/store-api-sdk/src/store_client.rs
@@ -21,8 +21,8 @@ use common_arrow::arrow_flight::BasicAuth;
 use common_arrow::arrow_flight::HandshakeRequest;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_store_api::util::STORE_RUNTIME;
-use common_store_api::util::STORE_SYNC_CALL_TIMEOUT;
+use common_kv_api_util::STORE_RUNTIME;
+use common_kv_api_util::STORE_SYNC_CALL_TIMEOUT;
 use common_tracing::tracing;
 use futures::stream;
 use futures::StreamExt;
