--- common/store-api/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "common-store-api"
-version = "0.1.0"
-authors = ["Databend Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-common-exception = {path = "../exception"}
-common-metatypes = {path= "../metatypes"}
-common-planners  = {path= "../planners"}
-common-datavalues= {path= "../datavalues"}
-common-streams = {path= "../streams"}
-common-datablocks = {path= "../datablocks"}
-common-runtime = { path = "../runtime"}
-common-tracing = { path = "../tracing"}
-common-store-api-util= { path = "../kv-apis/util" }
-common-kv-api= { path = "../kv-apis/api" }
-common-kv-api-vo= { path = "../kv-apis/vo" }
-
-serde = { version = "1.0", features = ["derive"] }
-futures = "0.3"
-async-trait = "0.1"
-thiserror = "1.0.29"
-lazy_static = "1.4.0"
-
-[dev-dependencies]
