+++ kvlocal/Cargo.toml
@@ -19,7 +19,7 @@ common-runtime = {path = "../common/runtime"}
 common-tracing = {path = "../common/tracing"}
 common-kv-api = {path = "../common/kv-apis/api" }
 common-kv-api-vo = {path = "../common/kv-apis/vo" }
-common-store-api-util = {path = "../common/kv-apis/util" }
+common-kv-api-util = {path = "../common/kv-apis/util" }
 common-sled-store = {path = "../common/sled-store"}
 common-store-api = {path = "../common/store-api"}
 common-tracing = {path = "../common/tracing"}
