+++ kvlocal/src/local_kv_store.rs
@@ -19,6 +19,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use common_exception::Result;
 use common_kv_api::KVApi;
+use common_kv_api_util::STORE_RUNTIME;
 use common_kv_api_vo::GetKVActionResult;
 use common_kv_api_vo::MGetKVActionResult;
 use common_kv_api_vo::PrefixListReply;
@@ -32,7 +33,6 @@ use common_raft_store::state_machine::AppliedState;
 use common_raft_store::state_machine::StateMachine;
 use common_runtime::tokio::sync::Mutex;
 pub use common_sled_store::init_temp_sled_db;
-use common_store_api_util::STORE_RUNTIME;
 use common_tracing::tracing;
 
 /// Local storage that provides the API defined by `KVApi`.
