+++ query/Cargo.toml
@@ -35,12 +35,13 @@ common-runtime = { path = "../common/runtime" }
 common-streams = { path = "../common/streams" }
 common-tracing = { path = "../common/tracing" }
 common-profling = { path = "../common/profiling" }
-common-store-api = { path = "../common/store-api" }
 common-store-api-sdk= {path = "../common/store-api-sdk" }
 common-kv-api = { path = "../common/kv-apis/api" }
 common-kv-api-vo = { path = "../common/kv-apis/vo" }
 common-meta-api = { path = "../common/meta-apis/api" }
 common-meta-api-vo = { path = "../common/meta-apis/vo" }
+common-dfs-api = { path = "../common/dfs-apis/api" }
+common-dfs-api-vo = { path = "../common/dfs-apis/vo" }
 common-io = { path = "../common/io" }
 common-metatypes = { path = "../common/metatypes" }
 common-clickhouse-srv = { path = "../common/clickhouse-srv" }
