+++ query/src/datasources/table/remote/remote_table_do_read.rs
@@ -13,11 +13,11 @@
 //  limitations under the License.
 //
 
+use common_dfs_api_vo::ReadAction;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_planners::PlanNode;
 use common_planners::ReadDataSourcePlan;
-use common_store_api::ReadAction;
 use common_streams::ProgressStream;
 use common_streams::SendableDataBlockStream;
 use futures::StreamExt;
