+++ store/src/api/rpc/flight_service_test.rs
@@ -18,6 +18,7 @@ use std::time::UNIX_EPOCH;
 
 use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
+use common_dfs_api::StorageApi;
 use common_kv_api::KVApi;
 use common_meta_api::MetaApi;
 use common_meta_api_vo::*;
@@ -30,7 +31,6 @@ use common_planners::DropDatabasePlan;
 use common_planners::DropTablePlan;
 use common_planners::ScanPlan;
 use common_runtime::tokio;
-use common_store_api_sdk::StorageApi;
 use common_store_api_sdk::StoreClient;
 use common_tracing::tracing;
 use pretty_assertions::assert_eq;
